Update patch set 19
Patch Set 19:

> Patch Set 19:
> 
> Is anyone actually using this plugin?  If not, I would suggest that we deprecate it so as not to waste effort maintaining it any more.

We are using it. Compared to reviewers-by-blame, it has the nice abilities to balance review requests and add reviewers with +2; but on the other hand it cannot be configured to ignore draft changes (on 2.14; not sure how it will work with private/WIP changes).

It would certainly be nice to maintain a single plugin which would support all use cases, but at the moment many features we depend on are missing from reviewers-by-blame plugin.

Patch-set: 19
CC: Gerrit User 1017568 <1017568@173816e5-2b9a-37c3-8a2e-48639d4f1153>